Revert "comm"
This reverts commit ddd7aafb4f6d31daae804db2064597dcf1a33047.

+++ Meetico/Meetico/src/main/java/tn/esprit/meetico/service/RequestServiceImpl.java
@@ -1,8 +1,8 @@
 package tn.esprit.meetico.service;
 
-import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import tn.esprit.meetico.entity.Request;
 import tn.esprit.meetico.entity.Status;
@@ -20,7 +20,7 @@ public class RequestServiceImpl implements IRequestService {
 	private UserRepository userRepository;
 
 	@Override
-	public Request createRequest(Request request) {
+	public ResponseEntity<String> createRequest(Request request) {
 		String firstName = request.getFirstName().replaceAll("\\s+", " ");
 		firstName = firstName.replaceFirst("\\s", "");
 		firstName = replaceLast(firstName, " ", "");
@@ -36,7 +36,8 @@ public class RequestServiceImpl implements IRequestService {
 		request.setLastName(lastName);
 		request.setSendTime(null);
 		request.setStatus(Status.UNSENT);
-		return requestRepository.save(request);
+		requestRepository.save(request);
+		return ResponseEntity.ok().body("Request created successfully");
 	}
 	
 	 private String replaceLast(String string, String find, String replace) {
@@ -59,41 +60,39 @@ public class RequestServiceImpl implements IRequestService {
 	 }
 
 	@Override
-	public Request updateRequest(Long requestId, Request updation) {
+	public ResponseEntity<String> updateRequest(Long requestId, Request updation) {
 		Request request = requestRepository.findById(requestId).orElse(null);
-		request.setEmail(Optional.ofNullable(updation.getEmail()).orElse(request.getEmail()));
-		request.setFirstName(Optional.ofNullable(updation.getFirstName()).orElse(request.getFirstName()));
-		request.setGender(Optional.ofNullable(updation.getGender()).orElse(request.getGender()));
-		request.setLastName(Optional.ofNullable(updation.getLastName()).orElse(request.getLastName()));
-		request.setNic(Optional.ofNullable(updation.getNic()).orElse(request.getNic()));
-		request.setPhoneNumber(Optional.ofNullable(updation.getPhoneNumber()).orElse(request.getPhoneNumber()));
-		request.setStatus(Optional.ofNullable(updation.getStatus()).orElse(request.getStatus()));
-		return requestRepository.save(request);
+		if (request != null) {
+			request.setEmail(Optional.ofNullable(updation.getEmail()).orElse(request.getEmail()));
+			request.setFirstName(
+					Optional.ofNullable(updation.getFirstName()).orElse(request.getFirstName()));
+			request.setLastName(Optional.ofNullable(updation.getLastName()).orElse(request.getLastName()));
+			requestRepository.save(request);
+			return ResponseEntity.ok().body("Request updated successfully.");
+		}
+		return ResponseEntity.badRequest().body("No correspondance.");
 	}
 	
 	@Override
-	public void deleteRequest(Long requestId) {
-		requestRepository.deleteById(requestId);
-	}
-	
-	@Override
-	public List<Request> retrieveAllRequests(Long userId) {
-		return requestRepository.findAllByUserId(userId);
+	public ResponseEntity<String> deleteRequest(Long requestId) {
+		Request request = requestRepository.findById(requestId).orElse(null);
+		if (request != null) {
+			requestRepository.deleteById(requestId);
+			return ResponseEntity.ok().body("Request deleted successfully.");
+		}
+		return ResponseEntity.badRequest().body("No correspondance.");
 	}
 
 	@Override
-	public User assignSenderToRequest(Long requestId, Long senderId) {
+	public ResponseEntity<String> assignSenderToRequest(Long senderId, Long requestId) {
 		Request request = requestRepository.findById(requestId).orElse(null);
 		User user = userRepository.findById(senderId).orElse(null);
-		request.setSender(user);
-		requestRepository.save(request);
-		return user;
-	}
-	
-	@Override
-	public List<Request> searchForRequests(String input) {
-		List<Request> requests = requestRepository.findAllByInput(input);
-		return requests;
+		if (request != null && user != null) {
+			request.setSender(user);
+			requestRepository.save(request);
+			return ResponseEntity.ok().body("Sender assigned successfully.");
+		}
+		return ResponseEntity.badRequest().body("No correspondance.");
 	}
 
 }
